requirements.txt
pyyaml
cloudpickle
scikit-learn
unidecode
regex
scipy
requests
tqdm
logzero
nltk
python-telegram-bot
